libevent test server Werror sign-compare
https://github.com/warmcat/libwebsockets/issues/1131

+++ test-apps/test-server-libevent.c
@@ -257,7 +257,7 @@ int main(int argc, char **argv)
 	}
 #endif
 
-	for (n = 0; n < ARRAY_SIZE(sigs); n++) {
+	for (n = 0; n < (int)ARRAY_SIZE(sigs); n++) {
 		signals[n] = evsignal_new(event_base_loop, sigs[n], signal_cb, event_base_loop);
 
 		evsignal_add(signals[n], NULL);
